refactor: improved function and variable names
