--- src/js/json/recipes.json
@@ -1,44 +1,0 @@
-{
-  "recipes": [
-    {
-      "title": "Panqueca Americana",
-      "image": "../../assets/panquecaA.png",
-      "id": "1"
-    },
-    {
-      "title": "Bolo de Morango",
-      "image": "../../assets/BoloM.png.png.png",
-      "id": "2"
-    },
-    {
-      "title": "Pizza Sabores",
-      "image": "../../assets/pizzaS.png",
-      "id": "3"
-    },
-    {
-      "title": "Ovo com Salada",
-      "image": "../../assets/OvoS.png",
-      "id": "4"
-    },
-    {
-      "title": "Misto Quente",
-      "image": "../../assets/MistoQ.png",
-      "id": "5"
-    },
-    {
-      "title": "Pão com Fruta e Mel",
-      "image": "../../assets/paoFeM.png",
-      "id": "6"
-    },
-    {
-      "title": "X-Egg",
-      "image": "../../assets/xE.png",
-      "id": "7"
-    },
-    {
-      "title": "Almondega com Rúcula",
-      "image": "../../assets/AlmondegaR.png",
-      "id": "8"
-    }
-  ]
-}
